# ai-service/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "基层医疗安全型预问诊AI服务",
    "description": "基层医疗安全型预问诊AI服务接口文档。\n\n**安全声明：本服务仅供教学演示，不提供真实诊断或替代医生。**\n\n知识库内容仅供教学参考，不构成诊断或治疗建议。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/triage/analyze": {
      "post": {
        "tags": [
          "分诊分析"
        ],
        "summary": "分诊分析",
        "description": "对合成教学病例进行分诊分析，复用多Agent安全审核流程。\n\n**安全声明：本接口仅供教学演示，不提供真实诊断或替代医生。**",
        "operationId": "analyze_triage_api_v1_triage_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SymptomInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "description": "症状名称",
            "examples": [
              "胸部不适"
            ]
          },
          "severity": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Severity",
            "description": "严重程度（0-10）",
            "default": 5,
            "examples": [
              7
            ]
          },
          "duration": {
            "type": "string",
            "maxLength": 100,
            "title": "Duration",
            "description": "持续时间",
            "default": "",
            "examples": [
              "30分钟"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SymptomInput",
        "description": "症状输入"
      },
      "TriageRequest": {
        "properties": {
          "case_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Case Id",
            "description": "病例标识（合成教学用）",
            "examples": [
              "synthetic-triage-001"
            ]
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 150.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age",
            "description": "年龄"
          },
          "symptoms": {
            "items": {
              "$ref": "#/components/schemas/SymptomInput"
            },
            "type": "array",
            "title": "Symptoms",
            "description": "症状列表"
          },
          "red_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Red Flags",
            "description": "红旗标识列表"
          },
          "free_text": {
            "type": "string",
            "maxLength": 5000,
            "title": "Free Text",
            "description": "自由文本描述（合成教学病例）",
            "default": ""
          },
          "model_suggested_risk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Suggested Risk",
            "description": "模型建议的风险等级"
          }
        },
        "type": "object",
        "required": [
          "case_id"
        ],
        "title": "TriageRequest",
        "description": "分诊分析请求"
      },
      "TriageResponse": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id",
            "description": "病例标识"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "请求追踪ID（UUID4）"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level",
            "description": "风险等级"
          },
          "symptom_summary": {
            "type": "string",
            "title": "Symptom Summary",
            "description": "症状摘要"
          },
          "red_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Red Flags",
            "description": "红旗标识列表"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence",
            "description": "证据列表"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations",
            "description": "引用来源列表"
          },
          "missing_information": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Information",
            "description": "缺失信息字段"
          },
          "followup_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Followup Questions",
            "description": "追问问题列表"
          },
          "safety_status": {
            "type": "string",
            "title": "Safety Status",
            "description": "安全状态：pass / blocked / human_review"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "description": "免责声明"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "trace_id",
          "risk_level",
          "symptom_summary",
          "red_flags",
          "evidence",
          "citations",
          "missing_information",
          "followup_questions",
          "safety_status",
          "disclaimer"
        ],
        "title": "TriageResponse",
        "description": "分诊分析响应"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}